lambda/requirements.txt
python-docx==1.1.2
requests==2.31.0
lxml==5.3.0
numpy==2.0.2
